+++ user-interface/src/main/java/life/qbic/datamanager/views/general/DragDropList.java
@@ -1,4 +1,4 @@
-package life.qbic.datamanager.views.demo;
+package life.qbic.datamanager.views.general;
 
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEventListener;
