+++ user-interface/src/main/java/life/qbic/datamanager/views/projects/project/experiments/experiment/components/ExperimentalVariablesInput.java
@@ -1,469 +1,580 @@
 package life.qbic.datamanager.views.projects.project.experiments.experiment.components;
 
-import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
-import static java.util.Objects.requireNonNull;
-import static life.qbic.datamanager.views.general.dialog.InputValidation.failed;
-import static life.qbic.datamanager.views.general.dialog.InputValidation.passed;
 
-import com.vaadin.flow.component.AbstractField;
+import com.vaadin.flow.component.AbstractField.ComponentValueChangeEvent;
 import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.ComponentEvent;
+import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.Composite;
 import com.vaadin.flow.component.DomEvent;
 import com.vaadin.flow.component.EventData;
-import com.vaadin.flow.component.HasValue;
+import com.vaadin.flow.component.Focusable;
+import com.vaadin.flow.component.HasValue.ValueChangeListener;
 import com.vaadin.flow.component.Key;
+import com.vaadin.flow.component.KeyModifier;
+import com.vaadin.flow.component.ShortcutRegistration;
+import com.vaadin.flow.component.Shortcuts;
+import com.vaadin.flow.component.Tag;
 import com.vaadin.flow.component.button.Button;
-import com.vaadin.flow.component.customfield.CustomField;
-import com.vaadin.flow.component.dnd.DragSource;
-import com.vaadin.flow.component.dnd.DropEffect;
-import com.vaadin.flow.component.dnd.DropTarget;
 import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.component.html.Span;
-import com.vaadin.flow.component.icon.Icon;
+import com.vaadin.flow.component.html.NativeLabel;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.progressbar.ProgressBar;
+import com.vaadin.flow.component.shared.HasValidationProperties;
 import com.vaadin.flow.component.textfield.TextField;
 import com.vaadin.flow.dom.DisabledUpdateMode;
+import com.vaadin.flow.function.SerializablePredicate;
 import com.vaadin.flow.shared.Registration;
-import java.io.Serial;
 import java.io.Serializable;
 import java.util.ArrayList;
-import java.util.HashSet;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.Optional;
-import java.util.function.Consumer;
+import java.util.Set;
+import java.util.StringJoiner;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import life.qbic.application.commons.CanSnapshot;
+import life.qbic.application.commons.Snapshot;
+import life.qbic.datamanager.views.general.ButtonFactory;
+import life.qbic.datamanager.views.general.DragDropList;
 import life.qbic.datamanager.views.general.dialog.InputValidation;
 import life.qbic.datamanager.views.general.dialog.UserInput;
-import life.qbic.datamanager.views.projects.project.experiments.experiment.components.ExperimentalVariablesInput.ExperimentalVariableRow.RemoveButtonClickEvent;
+import life.qbic.datamanager.views.projects.project.experiments.experiment.components.ExperimentalVariablesInput.VariableLevelsInput.LevelChange;
+import life.qbic.datamanager.views.projects.project.experiments.experiment.components.ExperimentalVariablesInput.VariableRow.InvalidChangesException;
+import life.qbic.datamanager.views.projects.project.experiments.experiment.components.ExperimentalVariablesInput.VariableRow.VariableChange;
+import life.qbic.datamanager.views.projects.project.experiments.experiment.components.ExperimentalVariablesInput.VariableRow.VariableDeleted;
 import org.springframework.lang.NonNull;
 import org.springframework.lang.Nullable;
 
-public class ExperimentalVariablesInput extends Div implements UserInput {
+public class ExperimentalVariablesInput extends Composite<Div> implements UserInput, CanSnapshot {
 
-
-  private final Div variableInformationContainer = new Div();
-  private List<ExperimentalVariableRow> initialState;
+  private final Button addVariableButton = new Button("Add Variable", VaadinIcon.PLUS.create());
+  private final Div variablesInput = new Div();
+  private final List<Snapshot> deletedVariables = new ArrayList<>();
+  private Snapshot initialState;
 
   public ExperimentalVariablesInput() {
-    initLayout();
-    markInitialState();
-    ensureAtLeastOneInputPresent();
+    addVariableButton.addClickListener(
+        e -> addVariable());
+    markInitialized();
+
+
+/*    //TEST CONTROLS FIXME REMOVE BEFORE MERGE
+    getContent().add(new Button("Add all deleted variables", e -> {
+      var copy = List.copyOf(deletedVariables);
+      copy.reversed() //FIFO
+          .forEach(deletedVariable -> {
+            addRow(toRestoredVariableRow(deletedVariable));
+            deletedVariables.remove(deletedVariable);
+          });
+    }));
+    TextArea inputField = new TextArea();
+    getContent().add(inputField);
+    getContent().add(new Button("Lock Levels", e -> {
+      inputField.getOptionalValue()
+          .map(it -> it.lines().collect(Collectors.toSet()))
+          .ifPresent(lines -> getVariableRows().forEach(row -> row.lockLevels(lines)));
+    }));
+    getContent().add(new Button("Avoid Deletion", e -> {
+      inputField.getOptionalValue()
+          .map(it -> it.lines().collect(Collectors.toSet()))
+          .ifPresent(lines -> getVariableRows()
+              .forEach(row -> row.setDeletable(!lines.contains(row.getVariableName()))));
+    }));*/
   }
 
-  private void initLayout() {
-    variableInformationContainer.addClassNames("flex-vertical");
-    add(variableInformationContainer);
-    var addVariableIcon = VaadinIcon.PLUS.create();
-    addVariableIcon.addClassNames("color-primary-text clickable");
-    addVariableIcon.addClickListener(it -> addNewVariableInput());
-    Span addVariableText = new Span("Add Experimental Variable");
-    addVariableText.addClassNames("color-primary-text clickable");
-    addVariableText.addClickListener(it -> addNewVariableInput());
-    var addVariableControl = new Div(addVariableIcon, addVariableText);
-    addVariableControl.addClassNames("flex-horizontal", "gap-04");
-    add(addVariableControl);
-  }
+  public record ExperimentalVariableInformation(String name, String unit, List<String> levels) {
 
-  private void markInitialState() {
-    initialState = rows().stream().filter(it -> !it.isEmpty()).toList();
-  }
-
-  public void editUnusedVariable(@NonNull ExperimentalVariableInformation variableInformation) {
-    addRow(new ExperimentalVariableRow(variableInformation));
-  }
-
-  public void editUsedVariable(@NonNull ExperimentalVariableInformation variableInformation,
-      @NonNull List<String> usedLevels) {
-    if (!(new HashSet<>(variableInformation.levels())).containsAll(usedLevels)) {
-      throw new IllegalArgumentException(
-          "The variable information must contain all used levelInputs");
-    }
-    addRow(new ExperimentalVariableRow(variableInformation));
-  }
-
-  public void addNewVariableInput() {
-    addRow(new ExperimentalVariableRow());
-  }
-
-  void deleteUnusedVariable(@NonNull String variableName) {
-    //TODO implement
-    throw new RuntimeException("Not implemented");
-  }
-
-
-  void lockUnit(@NonNull ExperimentalVariableRow row) {
-    //TODO implement
-    throw new RuntimeException("Not implemented");
-  }
-
-  void lockLevels(@NonNull ExperimentalVariableRow row, @NonNull String... levels) {
-    //TODO implement
-    throw new RuntimeException("Not implemented");
-  }
-
-  void setUnit(@NonNull String unit) {
-    //TODO implement
-    throw new RuntimeException("Not implemented");
   }
 
   public List<VariableChange> getChanges() {
-    //TODO implement compute changes and report
-    throw new RuntimeException("Not implemented");
+    return Stream.concat(
+        deletedVariables.stream()
+            .map(this::toRestoredVariableRow)
+            .map(VariableRow::getVariableName)
+            .map(VariableDeleted::new),
+        this.getVariableRows().stream()
+            .flatMap((VariableRow variableRow) -> {
+              try {
+                List<VariableChange> changes = variableRow.getChanges();
+                return changes.stream();
+              } catch (InvalidChangesException ex) {
+                return Stream.of();
+              }
+            })).toList();
   }
 
-  record VariableChange(
-      String oldVarName,
-      String newVarName,
-      String oldUnit,
-      String newUnit,
-      List<String> oldLevels,
-      List<String> newLevels) {
-
-    VariableChange {
-      boolean noOldVarName = isNull(oldVarName) || oldVarName.isBlank();
-      boolean noNewVarName = isNull(newVarName) || newVarName.isBlank();
-      if (noOldVarName && noNewVarName) {
-        throw new IllegalArgumentException("No variable name provided");
-      }
-    }
-
-    Optional<VariableRenaming> renaming() {
-      if (isVariableRenaming()) {
-        return Optional.of(new VariableRenaming(oldVarName, newVarName));
-      }
-      return Optional.empty();
-    }
-
-    Optional<VariableUnitChange> unitChange() {
-      if (isVariableUnitChange()) {
-        return Optional.of(new VariableUnitChange(oldUnit, newUnit));
-      }
-      return Optional.empty();
-    }
-
-    Optional<VariableCreation> creation() {
-      if (isVariableCreation()) {
-        return Optional.of(new VariableCreation(new ExperimentalVariableInformation(
-            newVarName,
-            newLevels,
-            newUnit
-        )));
-      }
-      return Optional.empty();
-    }
-
-    Optional<VariableDeletion> deletion() {
-      if (isVariableDeletion()) {
-        return Optional.of(new VariableDeletion(oldVarName));
-      }
-      return Optional.empty();
-    }
-
-    boolean isVariableRenaming() {
-      return isNull(oldVarName) || !oldVarName.equals(newVarName);
-    }
-
-    boolean isVariableCreation() {
-      return isNull(oldVarName) && !isNull(newVarName);
-    }
-
-    boolean isVariableUnitChange() {
-      if (nonNull(oldUnit)) {
-        if (isNull(newUnit)) {
-          return true;
-        }
-        return oldUnit.equals(newUnit);
-      }
-      return nonNull(newUnit);
-    }
-
-    boolean isVariableDeletion() {
-      return nonNull(oldVarName) && isNull(newVarName);
-    }
-
-    List<String> deletedLevels() {
-      return oldLevels.stream().filter(it -> !newLevels.contains(it)).toList();
-    }
-
-    List<String> addedLevels() {
-      return newLevels.stream().filter(it -> !oldLevels.contains(it)).toList();
-    }
-  }
-
-
-  record VariableRenaming(String oldName, String newName) {
-
-  }
-
-  record VariableCreation(ExperimentalVariableInformation variableInformation) {
-  }
-
-  record VariableDeletion(String variableName) {
-
-  }
-
-  record VariableUnitChange(String oldUnit, String newUnit) {
-
-  }
-
-  // when variable was created it cannot be deleted
-  // when variable was deleted no other changes to the variable apply
-  // when level was deleted no other changes to the level apply
-
-  // the parent needs to know which variables to create, which variables to delete, which variables to rename, which levels to create and which levels to remove
-  // unit changing is only allowed if none of the levels were in use which is only the case if no exp groups exist
-
-  // renaming must be done last or first, as the variable might not be found otherwise
-  // changes to one variable must happen in one flow as to not block each other.
-
-  /* so the caller must
-  1. query variable deletion -> make a snapshot for rollback and delete
-  2. query for variable level deletion -> make a snapshot for rollback and delete
-  3. query for variable unit change -> make a snapshot for rollback and delete
-  4. query for variable level addition -> make a snapshot for rollback and create
-  5. query for variable creation -> make a snapshot for rollback and create
-  6. query for variable renaming -> make a snapshot for rollback and rename
-   */
-
-  ExperimentalVariableRow getRow(ExperimentalVariableInformation experimentalVariableInformation) {
-    return rows().stream()
-        .filter(it -> experimentalVariableInformation.equals(it.initialValue))
-        .findAny().orElseThrow(() -> new IllegalArgumentException(
-            "Variable " + experimentalVariableInformation + " is not set in this dialog"));
-  }
-
-  private void addRow(ExperimentalVariableRow rowLayout) {
-    rowLayout.setRemoveButtonListener(it -> removeRow(rowLayout, it));
-    //TODO record change
-//    rowLayout.addValueChangeListener(event -> recordChange(rowLayout, event));
-    variableInformationContainer.addComponentAtIndex(rows().size(), rowLayout);
-  }
-
-  private void removeRow(ExperimentalVariableRow rowLayout, RemoveButtonClickEvent it) {
-    variableInformationContainer.remove(it.origin());
-    ensureAtLeastOneInputPresent();
-  }
-
-  private void ensureAtLeastOneInputPresent() {
-    if (rows().isEmpty()) {
-      //if there are no rows after the deletion, create an empty one.
-      addNewVariableInput();
-    }
-  }
-
-  private List<ExperimentalVariableRow> rows() {
-    return variableInformationContainer
-        .getChildren()
-        .filter(ExperimentalVariableRow.class::isInstance)
-        .map(ExperimentalVariableRow.class::cast)
-        .toList();
+  public void markInitialized() {
+    getVariableRows().forEach(VariableRow::markInitialized);
+    this.deletedVariables.clear();
+    this.initialState = this.snapshot();
   }
 
   @NonNull
+  private List<VariableRow> getVariableRows() {
+    return variablesInput.getChildren()
+        .filter(VariableRow.class::isInstance)
+        .map(VariableRow.class::cast)
+        .toList();
+  }
+
   @Override
+  @NonNull
+  protected Div initContent() {
+    var body = new Div();
+    body.addClassNames("dialog-section");
+    variablesInput.addClassNames("flex-vertical gap-04");
+
+    addVariable();
+
+    addVariableButton.addClassNames(
+        "margin-bottom-04 flex-vertical width-max-content justify-self-start button-color-primary");
+    variablesInput.add(addVariableButton);
+    body.add(variablesInput);
+    return body;
+  }
+
+  public void addVariable() {
+    VariableRow variableRow = new VariableRow();
+    addRow(variableRow);
+  }
+
+  public void addVariable(ExperimentalVariableInformation variableInformation) {
+    if (!getVariableRows().isEmpty()) {
+      VariableRow firstRow = getVariableRows().getFirst();
+      if (firstRow != null && firstRow.isEmpty()) {
+        firstRow.removeFromParent();
+      }
+    }
+
+    VariableRow variableRow = new VariableRow();
+    variableRow.setValue(variableInformation);
+    variableRow.markInitialized();
+    addRow(variableRow);
+  }
+
+  private void addRow(VariableRow variableRow) {
+    variableRow.addDeleteVariableListener(it -> removeVariable(it.getSource()));
+    variablesInput.addComponentAtIndex(
+        Math.max(0, (int) (variablesInput.getChildren().count() - 1)),
+        variableRow);
+    variableRow.focus();
+  }
+
+  private void removeVariable(@NonNull VariableRow variableRow) {
+    if (!variableRow.isEmpty()) {
+      var initialComponentState = toRestoredVariablesInput(this.initialState);
+      var initialRowInformation = toRestoredVariableRow(variableRow.initialState);
+      if (initialComponentState.getVariableRows().stream()
+          .anyMatch(row -> row.getVariableName().equals(initialRowInformation.getVariableName()))) {
+        //only record deletion of variable rows that were there initially
+        var snapshotOfVariableRow = variableRow.snapshot();
+        deletedVariables.add(snapshotOfVariableRow);
+      }
+    }
+    variableRow.removeFromParent();
+    if (getVariableRows().isEmpty()) {
+      addVariable();
+    }
+  }
+
+  private ExperimentalVariablesInput toRestoredVariablesInput(Snapshot snapshot) {
+    var dummy = new ExperimentalVariablesInput();
+    try {
+      dummy.restore(snapshot);
+      return dummy;
+    } catch (SnapshotRestorationException e) {
+      throw new IllegalStateException("Initial state not correctly set.", e);
+    }
+  }
+
+
+  @Override
+  @NonNull
   public InputValidation validate() {
-    List<ExperimentalVariableRow> rows = rows();
-    if (rows.isEmpty()) {
-      return passed();
-    }
-    if (rows.stream()
-        .map(ExperimentalVariableRow::validate)
-        .allMatch(InputValidation::hasPassed)) {
-      return passed();
-    }
-    return failed();
+    //the order of the calls matters as both methods display the error.
+    var individualRowValidations = validateIndividualRows();
+    var overallRulesValidation = validateOverallRules();
+    return individualRowValidations.and(overallRulesValidation);
+
+  }
+
+  /**
+   * Checks rules that cannot be checked on an individual row basis. Modifies the userinput to
+   * communicate the validation result.
+   *
+   * @return the resulting validation
+   */
+  private InputValidation validateOverallRules() {
+    return checkUniqueVariableNames();
+  }
+
+  private InputValidation checkUniqueVariableNames() {
+    List<VariableRow> variableRows = getVariableRows();
+    Map<String, Long> countsByName = variableRows.stream()
+        .collect(Collectors.groupingBy(VariableRow::getVariableName, Collectors.counting()));
+
+    List<VariableRow> rowsWithDuplicateName = variableRows.stream()
+        .filter(it -> !it.isEmpty())
+        .filter(it -> !it.getVariableName().isBlank())
+        .filter(row -> countsByName.getOrDefault(row.getVariableName(), 0L) > 1)
+        .toList();
+    rowsWithDuplicateName.forEach(row -> {
+      // be careful not to overwrite previously set invalid state as marking as valid is done on an individual level only.
+      row.setNameInvalid("This variable name already exists.");
+    });
+    return !rowsWithDuplicateName.isEmpty() ? InputValidation.failed() : InputValidation.passed();
+  }
+
+  /**
+   * Triggers the validation of all variable rows.
+   *
+   * @return the resulting validation
+   */
+  private InputValidation validateIndividualRows() {
+    //we cannot use Stream#allMatch as this will not guarantee a call to VariableRow#validate on each row.
+    return getVariableRows().stream()
+        .map(VariableRow::validate)
+        .reduce(InputValidation.passed(), InputValidation::and);
   }
 
   @Override
   public boolean hasChanges() {
-    List<ExperimentalVariableRow> rows = rows();
-    if (initialState.isEmpty() && rows.isEmpty()) {
-      return false;
+    return !deletedVariables.isEmpty() || getVariableRows().stream()
+        .anyMatch(VariableRow::hasChanges);
+  }
+
+  record VariablesSnapshot(List<Snapshot> variableSnapshots,
+                           @Nullable Snapshot initialState) implements Snapshot {
+
+    public VariablesSnapshot {
+      variableSnapshots = variableSnapshots.stream().toList();
     }
-    if (initialState.size() != getChanges().size()) {
-      // added or removed something
-      return true;
-    } else {
-      //same length
-      if (rows.stream().anyMatch(ExperimentalVariableRow::hasChanges)) {
-        return true;
+
+  }
+
+  @Override
+  public Snapshot snapshot() {
+    return new VariablesSnapshot(getVariableRows().stream().map(VariableRow::snapshot).toList(),
+        initialState);
+  }
+
+  private VariableRow toRestoredVariableRow(Snapshot snapshot) {
+    var dummy = new VariableRow();
+    try {
+      dummy.restore(snapshot);
+      return dummy;
+    } catch (SnapshotRestorationException e) {
+      throw new IllegalStateException("Initial state not correctly set.", e);
+    }
+  }
+
+  @Override
+  public void restore(Snapshot snapshot) throws SnapshotRestorationException {
+    if (snapshot instanceof VariablesSnapshot(List<Snapshot> variableSnapshots, Snapshot state)) {
+      if (nonNull(state)
+          && state instanceof VariablesSnapshot) {
+        this.initialState = state;
       }
-      return !new HashSet<>(rows.stream()
-          .map(HasValue::getOptionalValue).toList())
-          .containsAll(initialState.stream()
-              .map(ExperimentalVariableRow::getOptionalValue).toList());
+      var previousChildren = getVariableRows();
+      previousChildren.forEach(Component::removeFromParent);
+      try {
+        variableSnapshots
+            .stream()
+            .map(this::toRestoredVariableRow)
+            .forEach(this::addRow);
+        return;
+      } catch (Exception e) {
+        getVariableRows().forEach(Component::removeFromParent);
+        previousChildren.forEach(this::addRow);
+        throw e;
+      }
     }
+    throw new SnapshotRestorationException(
+        "Unknown snapshot type. Expected %s but got %s".formatted(VariablesSnapshot.class,
+            snapshot.getClass()));
   }
 
-  public record ExperimentalVariableInformation(String variableName, List<String> levels,
-                                                @Nullable String unit) implements Serializable {
 
-    public ExperimentalVariableInformation {
-      requireNonNull(variableName);
-      requireNonNull(levels);
-      levels = List.copyOf(levels);
-    }
-  }
+  static class VariableRow extends Composite<Div> implements UserInput, CanSnapshot {
 
-  static final class LevelArea extends Composite<Div> {
+    private final TextField name = new TextField();
+    private final TextField unit = new TextField();
+    private final VariableLevelsInput variableLevels = VariableLevelsInput.empty();
+    private final Button deleteVariable = new ButtonFactory()
+        .createTertirayButton("Delete Variable", VaadinIcon.TRASH.create());
+    private Snapshot initialState;
 
-    private Div helperText;
-    private Div levelContainer;
-    private final List<TextField> levels = new ArrayList<>();
-    private final List<DropTarget<?>> dropTargets = new ArrayList<>();
-    private Component helper;
-    private Div errorText;
-
-    public LevelArea() {
-      var content = getContent();
-      helperText = new Div();
-      helperText.setVisible(false);
-      levelContainer = new Div();
-      levelContainer.addClassNames("flex-vertical");
-      errorText = new Div("Please provide at least one level.");
-      errorText.setVisible(false);
-      var addLevelControls = createAddLevelControls();
-      content.add(levelContainer, addLevelControls, errorText, helperText);
-      addLevelPasteListener();
+    public VariableRow() {
+      deleteVariable.addClickListener(
+          e -> fireEvent(new DeleteVariableEvent(this, e.isFromClient())));
+      markInitialized();
     }
 
-    public void setRequired(boolean required) {
-      getContent().getElement().setProperty("required", required);
+    public void setValue(ExperimentalVariableInformation variableInformation) {
+      name.setValue(variableInformation.name());
+      unit.setValue(variableInformation.unit());
+      variableInformation.levels().forEach(variableLevels::addFilledLevel);
     }
 
-    public boolean isRequired() {
-      return getContent().getElement().getProperty("required", false);
+    public void markInitialized() {
+      variableLevels.markInitialized();
+      this.initialState = snapshot();
+    }
+
+    public Registration addDeleteVariableListener(
+        ComponentEventListener<DeleteVariableEvent> listener) {
+      return addListener(DeleteVariableEvent.class, listener);
+    }
+
+
+    @Override
+    protected Div initContent() {
+      var root = new Div();
+      root.addClassNames(
+          "border rounded-02 padding-04 gap-04 column-gap-05 grid-experimental-variable-input");
+      var fields = new Div();
+      fields.getStyle().set("grid-area", "a");
+      fields.addClassNames("flex-horizontal gap-05");
+      name.addClassNames("dynamic-growing-flex-item");
+      name.setLabel("Variable Name");
+      name.setRequired(true);
+      unit.addClassNames("dynamic-growing-flex-item");
+      unit.setLabel("Unit (optional)");
+      unit.setRequired(false);
+      variableLevels.getStyle().set("grid-area", "b");
+      deleteVariable.getStyle().set("grid-area", "c");
+      deleteVariable.addClassNames("width-max-content");
+      fields.add(name, unit);
+      root.add(fields, variableLevels, deleteVariable);
+      return root;
+    }
+
+    public void setDeletable(boolean deletable) {
+      deleteVariable.setEnabled(deletable);
+      deleteVariable.setVisible(deletable);
+    }
+
+    /**
+     * locked levels can be reordered but not deleted or renamed
+     *
+     * @param levels
+     */
+    public void lockLevels(Set<String> levels) {
+      variableLevels.lockLevels(levels);
+      if (variableLevels.getLevels().stream().anyMatch(levels::contains)) {
+        setDeletable(false);
+        lock(unit);
+      }
+    }
+
+    private void lock(TextField textField) {
+      textField.setSuffixComponent(VaadinIcon.LOCK.create());
+      textField.setReadOnly(true);
+    }
+
+    @Override
+    @NonNull
+    public InputValidation validate() {
+      if (this.isEmpty()) {
+        setInvalid(false);
+        return InputValidation.passed();
+      }
+
+      InputValidation nameValidation;
+      if (name.getOptionalValue().map(String::isBlank).orElse(true)) {
+        setChildValid(name);
+        setNameInvalid("Please provide a variable name.");
+        nameValidation = InputValidation.failed();
+      } else {
+        setChildValid(name);
+        nameValidation = InputValidation.passed();
+      }
+
+      InputValidation unitValidation = InputValidation.passed();
+      InputValidation variableLevelsValidation = variableLevels.validate();
+      return nameValidation
+          .and(unitValidation)
+          .and(variableLevelsValidation);
     }
 
     public void setInvalid(boolean invalid) {
-      getContent().getElement().setProperty("invalid", invalid);
-      errorText.setVisible(invalid);
+      getElement().setProperty("invalid", invalid);
+      getElement().setAttribute("invalid", invalid);
+      if (!invalid) {
+        setChildrenValid();
+      }
     }
 
+    private void setChildrenValid() {
+      setChildValid(name);
+      setChildValid(unit);
+      setChildValid(variableLevels);
+    }
 
-    private TextField createAddLevelControls() {
-      TextField textField = new TextField();
-      var addBtn = new Button(VaadinIcon.PLUS.create());
-      addBtn.setText("Add");
-      addBtn.setAriaLabel("Add Level");
-      addBtn.addClickListener(clickEvent -> {
-        var value = textField.getOptionalValue();
-        if (value.map(it -> !it.isBlank()).isEmpty()) {
-          return;
+    private void setChildValid(HasValidationProperties name) {
+      name.setErrorMessage(null);
+      name.setInvalid(false);
+    }
+
+    public void setNameInvalid(@NonNull String errorMessage) {
+      this.setInvalid(true);
+      if (Optional.ofNullable(name.getErrorMessage()).map(it -> !it.contains(errorMessage)).orElse(
+          true)) {
+        String composedErrorMessage = Optional.ofNullable(name.getErrorMessage())
+            .map(m -> m + "\n" + errorMessage)
+            .orElse(errorMessage);
+        name.setErrorMessage(composedErrorMessage);
+      }
+      name.setInvalid(true);
+    }
+
+    public void setUnitInvalid(@NonNull String errorMessage) {
+      this.setInvalid(true);
+      unit.setErrorMessage(errorMessage);
+      unit.setInvalid(true);
+    }
+
+    private boolean isEmpty() {
+      return name.isEmpty() && unit.isEmpty() && variableLevels.isEmpty();
+    }
+
+    private String getVariableName() {
+      return name.getValue();
+    }
+
+    private Optional<String> getUnit() {
+      return unit.getOptionalValue();
+    }
+
+    private List<LevelChange> getLevelChanges() {
+      return variableLevels.getChanges();
+    }
+
+    public static class InvalidChangesException extends RuntimeException {
+
+
+    }
+
+    /**
+     * Validates the component and returns all changes. Throws InvalidChangesException in case of
+     * invalid input.
+     *
+     * @return a list of variable changes
+     * @throws InvalidChangesException in case of invalid user input
+     */
+    public List<VariableChange> getChanges() throws InvalidChangesException {
+      if (!hasChanges()) {
+        return List.of(); //avoid unnecessary work
+      }
+      if (!validate().hasPassed()) {
+        throw new InvalidChangesException();
+      }
+      if (initialState instanceof ExperimentalVariableSnapshot experimentalVariableSnapshot) {
+        var previousRepresentation = new VariableRow();
+        previousRepresentation.restore(experimentalVariableSnapshot);
+        var changes = new ArrayList<VariableChange>();
+        if (previousRepresentation.getVariableName().isBlank() && !this.getVariableName()
+            .isBlank()) {
+          return Collections.singletonList(
+              new VariableAdded(getVariableName(), getUnit().orElse(null),
+                  variableLevels.getLevels()));
         }
-        addLevel(textField.getValue());
-        textField.clear();
-        textField.focus();
-      });
-      addBtn.addClickShortcut(Key.ENTER);
-      textField.setPlaceholder("Enter a new level value");
-      textField.setSuffixComponent(addBtn);
-      return textField;
+        if (!previousRepresentation.getVariableName().equals(this.getVariableName())) {
+          changes.add(new VariableRenamed(previousRepresentation.getVariableName(),
+              this.getVariableName()));
+        }
+        if (!previousRepresentation.getUnit().equals(this.getUnit())) {
+          changes.add(new VariableUnitChanged(this.getVariableName(),
+              previousRepresentation.getUnit().orElse(null),
+              this.getUnit().orElse(null)));
+        }
+        if (!this.getLevelChanges().isEmpty()) {
+          changes.add(new VariableLevelsChanged(getVariableName(), this.getLevelChanges()));
+        }
+        return changes;
+      }
+      throw new IllegalStateException();
     }
 
-    Optional<String> getHelperText() {
-      return Optional.ofNullable(helperText.getText())
-          .filter(it -> !it.isBlank());
+    record ExperimentalVariableSnapshot(String name, String unit,
+                                        Snapshot levels, @Nullable Snapshot initialState) implements
+        Snapshot, Serializable {
+
+
     }
 
-    void addLevel(@NonNull String value) {
-      requireNonNull(value);
-      if (value.isBlank()) {
+    @Override
+    public boolean hasChanges() {
+      if (initialState == null) {
+        return !isEmpty();
+      }
+      if (initialState instanceof ExperimentalVariableSnapshot initialVariableState) {
+        return !initialVariableState.name().equals(getVariableName())
+            || !initialVariableState.unit().equals(unit.getValue())
+            || variableLevels.hasChanges();
+      }
+      throw new IllegalStateException("Could not check for changes of " + this);
+    }
+
+    @Override
+    public Snapshot snapshot() {
+      return new ExperimentalVariableSnapshot(name.getValue(), unit.getValue(),
+          variableLevels.snapshot(), initialState);
+    }
+
+    @Override
+    public void restore(@NonNull Snapshot snapshot) throws SnapshotRestorationException {
+      if (snapshot instanceof ExperimentalVariableSnapshot(
+          String name1, String unit1, Snapshot levels, Snapshot state
+      )) {
+        if (nonNull(state)) {
+          initialState = state;
+        }
+        variableLevels.restore(levels);
+        name.setValue(name1);
+        unit.setValue(unit1);
         return;
       }
-      TextField textField = new TextField();
-      textField.setValue(value);
-      var deleteBtn = VaadinIcon.TRASH.create();
-      deleteBtn.addClassName("clickable");
-      deleteBtn.addClickListener(it -> {
-        levels.remove(textField);
-        textField.removeFromParent();
-      });
-      var dragSource = DragSource.create(textField);
-      dragSource.setDraggable(true);
-      final String draggableCssClass = "draggable";
-      final String draggingInProgressCssClass = "dragging";
-      var dragHandle = VaadinIcon.ELLIPSIS_V.create();
-      dragHandle.addClassName(draggableCssClass);
-      dragSource.addDragStartListener(it -> {
-        //only allow drop in same variable
-        dropTargets.forEach(target -> target.setActive(true));
-        dragHandle.addClassName(draggingInProgressCssClass);
-        dragHandle.removeClassName(draggableCssClass);
-      });
-      dragSource.addDragEndListener(it -> {
-        //disable drop in this variable after drag end
-        dropTargets.forEach(target -> target.setActive(false));
-        dragHandle.removeClassName(draggingInProgressCssClass);
-        dragHandle.addClassName(draggableCssClass);
-      });
-      var dropTarget = DropTarget.configure(textField, false);
-      textField.addClassName("drop-above");
-      dropTarget.setDropEffect(DropEffect.MOVE);
-      dropTarget.addDropListener(it -> {
-        if (it.getDragSourceComponent().isEmpty()) {
-          return;
-        }
-        var dragSourceComponent = it.getDragSourceComponent().orElseThrow();
-        if (dragSourceComponent instanceof TextField sourceInput) {
-          int targetIndex = levelContainer.indexOf(it.getComponent());
-          int sourceIndex = levelContainer.indexOf(sourceInput);
-          if (targetIndex > -1 && sourceIndex > -1) {
-            levelContainer.addComponentAtIndex(targetIndex, sourceInput);
-          }
-        }
-      });
-      textField.setPrefixComponent(dragHandle);
-      textField.setSuffixComponent(deleteBtn);
-      levelContainer.add(textField);
-      dropTargets.add(dropTarget);
-      levels.add(textField);
+      throw new SnapshotRestorationException("Cannot restore snapshot");
     }
 
-    void setLevels(@NonNull List<String> values) {
-      levels.forEach(Component::removeFromParent);
-      levels.clear();
-      values.forEach(this::addLevel);
+    public void focus() {
+      name.focus();
     }
 
-    void setHelperText(@Nullable String text) {
-      this.helperText.setText(text);
-      this.helperText.setVisible(getHelperText().isPresent());
+    public interface VariableChange {
+
     }
 
-    List<String> getLevels() {
-      return levels.stream()
-          .map(TextField::getOptionalValue)
-          .filter(Optional::isPresent)
-          .map(Optional::get)
-          .toList();
+    public record VariableAdded(String name, String unit, List<String> levels) implements
+        VariableChange {
+
     }
 
-    private void addLevelPasteListener() {
-      addListener(PasteEvent.class, it -> {
-        List<String> pastedLines = it.getClipboardText().lines()
-            .filter(line -> !line.isBlank())
-            .toList();
-        pastedLines.forEach(this::addLevel);
-      });
+    public record VariableRenamed(String oldName, String newName) implements VariableChange {
+
     }
 
-    public boolean isEmpty() {
-      return levels.stream().map(HasValue::getOptionalValue)
-          .filter(Optional::isPresent)
-          .map(Optional::get)
-          .allMatch(String::isBlank);
+    public record VariableLevelsChanged(String name,
+                                        List<LevelChange> levelChanges) implements VariableChange {
+
     }
 
-    @DomEvent(value = "paste",
-        allowUpdates = DisabledUpdateMode.ONLY_WHEN_ENABLED,
-        preventDefault = true,
-        stopPropagation = true)
-    public static class PasteEvent extends ComponentEvent<Component> {
+    public record VariableUnitChanged(String name, String oldUnit, String newUnit) implements
+        VariableChange {
 
-      private final String clipboardText;
+    }
+
+    public record VariableDeleted(String name) implements VariableChange {
+
+    }
+
+    static class DeleteVariableEvent extends ComponentEvent<VariableRow> {
 
       /**
        * Creates a new event using the given source and indicator whether the event originated from
@@ -473,171 +584,550 @@ public class ExperimentalVariablesInput extends Div implements UserInput {
        * @param fromClient <code>true</code> if the event originated from the client
        *                   side, <code>false</code> otherwise
        */
-      public PasteEvent(Component source, boolean fromClient,
-          @EventData("event.clipboardData.getData(\"text\")") String clipboardText) {
+      public DeleteVariableEvent(VariableRow source, boolean fromClient) {
         super(source, fromClient);
-        this.clipboardText = clipboardText;
-      }
-
-      public String getClipboardText() {
-        return clipboardText;
       }
     }
   }
 
-  /**
-   * A layout containing rows for experimental variable input
-   */
-  static final class ExperimentalVariableRow extends
-      CustomField<ExperimentalVariableInformation> implements UserInput {
 
-    @Serial
-    private static final long serialVersionUID = 8301389606566298879L;
-    private final TextField nameField = new TextField("Experimental Variable");
-    private final TextField unitField = new TextField("Unit");
-    private final LevelArea levelArea = new LevelArea();
-    private final Icon deleteIcon = new Icon(VaadinIcon.CLOSE_SMALL);
-    private Registration clickListener;
-    private ExperimentalVariableInformation initialValue;
-    private final Div fieldsContainer = new Div();
-    private final List<AbstractField<?, ?>> fields = List.of(nameField, unitField);
-    private final ProgressBar progressBar;
+  @Tag("variable-levels-input")
+  static class VariableLevelsInput extends Div implements UserInput, CanSnapshot,
+      HasValidationProperties {
 
-    {
-      progressBar = new ProgressBar();
-      progressBar.setIndeterminate(true);
-      progressBar.setVisible(false);
+    private final DragDropList<LevelField> levelsContainer = new DragDropList<>(LevelField.class);
+    private Snapshot initialState;
+    LevelField focusedLevelField = null;
+
+    public List<String> getLevels() {
+      return filledLevels();
     }
 
-    ExperimentalVariableRow(ExperimentalVariableInformation initialValue) {
-      super(null);
-      init();
-      setValue(initialValue);
+    public void lockLevels(Set<String> levels) {
+      levelsContainer.stream()
+          .filter(field -> !field.isEmpty())
+          .filter(field -> levels.contains(field.getValue().orElseThrow()))
+          .forEach(LevelField::lock);
     }
 
-    ExperimentalVariableRow() {
-      super(null);
-      init();
-      setValue(getEmptyValue());
+
+    public interface LevelChange {
+
     }
 
-    private void init() {
-      fieldsContainer.addClassNames("flex-horizontal", "column-gap-05",
-          "flex-align-items-baseline");
-      nameField.setRequired(true);
-      nameField.setPlaceholder("e.g. age");
-      unitField.setPlaceholder("e.g. years");
-      levelArea.setRequired(true);
-      levelArea.setHelperText(
-          "Please enter each level on a new line. Comma separated values are treated as a single level.");
-      deleteIcon.addClassNames("color-primary-text");
-      /*Span around Icon is necessary otherwise icon size will be scaled down if a scrollbar appears*/
-      fieldsContainer.add(nameField, unitField, levelArea, new Span(deleteIcon));
-      add(fieldsContainer);
-      add(progressBar);
-      //TODO mark pending or completed for async loading
+    public record LevelAdded(int position, String value) implements LevelChange {
+
     }
 
+    public record LevelDeleted(int position) implements LevelChange {
+
+    }
+
+    public record LevelMoved(int oldPosition, int newPosition) implements LevelChange {
+
+    }
+
+    @NonNull
+    private LevelField addEmptyLevel() {
+      LevelField levelField = new LevelField();
+      addLevel(levelField);
+      levelField.focus();
+      return levelField;
+    }
+
+    @NonNull
+    private LevelField addFilledLevel(@NonNull String value) {
+      LevelField levelField = new LevelField();
+      levelField.setValue(value);
+      addLevel(levelField);
+      levelField.focus();
+      return levelField;
+    }
+
+    @NonNull
+    private LevelField addFilledLevel(int index, @NonNull String value) {
+      LevelField levelField = new LevelField();
+      levelField.setValue(value);
+      addLevel(index, levelField);
+      levelField.focus();
+      return levelField;
+    }
+
+
+    private void addLevel(LevelField levelField) {
+      addLevel(levelsContainer.size(), levelField);
+      afterAdd(levelField);
+    }
 
     /**
-     * @param value the new value. Can be null.
+     * Tries to add the level at the specified index. If the index is smaller than or equal to 0,
+     * the level is added as first level. If the index is greater than the current number of levels,
+     * the level is added as the last level.
+     *
+     * @param index      the index where the level field should be added
+     * @param levelField the field to add
      */
-    @Override
-    public void setValue(ExperimentalVariableInformation value) {
-      super.setValue(value);
-      this.initialValue = value;
+    private void addLevel(int index, LevelField levelField) {
+      levelsContainer.add(index, levelField);
+      afterAdd(levelField);
     }
 
-    @Override
-    public void setEnabled(boolean enabled) {
-      super.setEnabled(enabled);
-      fields.forEach(field -> field.setEnabled(enabled));
-      deleteIcon.getElement().setEnabled(enabled);
-      deleteIcon.setVisible(enabled);
+    private void afterAdd(LevelField levelField) {
+      levelField.addDeleteListener(event -> removeLevelField(event.getSource()));
+      // when the value changes, the validation is outdated and removed.
+      levelField.addValueChangeListener(valueChangeEvent -> setInvalid(false));
+      // when an empty field is added, the validation is outdated and removed.
+      setInvalid(false);
+      levelField.addFocusListener(event -> this.focusedLevelField = levelField);
+      levelField.addBlurListener(event -> this.focusedLevelField = null);
+      levelField.focus();
     }
 
-    @Override
-    protected ExperimentalVariableInformation generateModelValue() {
-      List<String> levels = levelArea.getLevels();
-      if (nameField.isEmpty() && levels.isEmpty() && unitField.isEmpty()) {
-        return null;
+    private void removeLevelField(LevelField levelField) {
+      levelsContainer.remove(levelField);
+      if (levelsContainer.isEmpty()) {
+        addEmptyLevel();
       }
-      return new ExperimentalVariableInformation(nameField.getValue(),
-          levels,
-          readUnitValue().orElse(null));
     }
 
-    private void setUnitValue(@Nullable String unit) {
-      if (isNull(unit)) {
-        this.unitField.clear();
+    public static VariableLevelsInput empty() {
+      var input = new VariableLevelsInput();
+      input.addEmptyLevel();
+      input.markInitialized();
+      return input;
+    }
+
+    public static VariableLevelsInput filled(@NonNull List<String> levels) {
+      if (levels.isEmpty()) {
+        return empty();
+      }
+      var input = new VariableLevelsInput();
+      levels.forEach(input::addFilledLevel);
+      input.markInitialized();
+      return input;
+    }
+
+
+    private VariableLevelsInput() {
+      final String rootCssClasses = "border rounded-02 flex-vertical gap-none padding-04 padding-top-04";
+      final String bodyClassNames = "flex-vertical justify-start gap-04 input-with-label";
+      this.getStyle().set("grid-area", "b");
+      this.addClassNames(rootCssClasses);
+
+      var body = new Div();
+      body.addClassNames(bodyClassNames);
+
+      final String levelsContainerCss = "flex-horizontal gap-03 column-gap-03 width-full";
+      levelsContainer.setId("levels-container"); //needed for labelling
+      levelsContainer.addClassNames(levelsContainerCss);
+      ButtonFactory buttonFactory = new ButtonFactory();
+
+      String labelCss = "form-label input-label";
+      var label = new NativeLabel("Levels");
+      label.setFor(levelsContainer);
+      label.addClassNames(labelCss);
+
+      var addLevelButton = buttonFactory.createTertirayButton("Add Level",
+          VaadinIcon.PLUS.create());
+      addLevelButton.addClickListener(clickEvent -> addEmptyLevel());
+
+      body.add(levelsContainer, addLevelButton);
+      addLevelButton.addClassNames("width-max-content justify-self-start");
+      this.add(label, body);
+
+      markInitialized();
+      Shortcuts.addShortcutListener(this, it -> {
+        var focusIndex = focusedLevelIndex();
+        int nextFocus = focusIndex + 1;
+        if (levelsContainer.size() > nextFocus) {
+          levelsContainer.getAt(nextFocus).focus();
+        } else {
+          addEmptyLevel();
+        }
+      }, Key.ENTER).listenOn(this);
+
+      addListener(PasteEvent.class, pasteEvent -> {
+        List<String> pastedLines = pasteEvent.getClipboardText().lines()
+            .filter(Objects::nonNull)
+            .filter(line -> !line.isBlank())
+            .toList();
+        if (pastedLines.isEmpty()) {
+          return;
+        }
+        var startIndex = focusedLevelIndex();
+        if (nonNull(focusedLevelField) && focusedLevelField.isEmpty()) {
+          //fill the first item in the focused field
+          focusedLevelField.setValue(pastedLines.getFirst());
+        } else {
+          startIndex++;
+          //fill the first item in a new field after the focused field
+          addFilledLevel(startIndex, pastedLines.getFirst());
+        }
+        for (int lineIndex = 1; lineIndex < pastedLines.size(); lineIndex++) {
+          var line = pastedLines.get(lineIndex);
+          var insertIndex = startIndex + lineIndex;
+          addFilledLevel(insertIndex, line);
+        }
+      });
+    }
+
+    private int focusedLevelIndex() {
+      return Optional.ofNullable(focusedLevelField)
+          .map(levelsContainer::indexOf)
+          .orElse(-1); //null leads to same result as indexOf if not a child yet
+    }
+
+
+    @Override
+    public void setInvalid(boolean invalid) {
+      HasValidationProperties.super.setInvalid(invalid);
+      getElement().setAttribute("invalid", invalid);
+      if (!invalid) {
+        levelsContainer.stream().forEach(f -> f.setErrorMessage(null));
+        levelsContainer.stream().forEach(f -> f.setInvalid(false));
+      } else {
+        levelsContainer.stream()
+            .findFirst()
+            .ifPresent(levelField -> {
+              levelField.setErrorMessage(getErrorMessage());
+              levelField.setInvalid(true);
+            });
+      }
+    }
+
+    Function<List<LevelField>, List<String>> extractFilledLevels() {
+      return fields -> fields.stream()
+          .map(LevelField::getValue)
+          .filter(Optional::isPresent)
+          .map(Optional::get)
+          .filter(s -> !s.isBlank())
+          .toList();
+    }
+
+    private List<String> filledLevels() {
+      return extractFilledLevels().apply(levelsContainer.stream().toList());
+    }
+
+    private boolean isEmpty() {
+      return filledLevels().isEmpty();
+    }
+
+    private record LevelsSnapshot(List<Snapshot> levels) implements Snapshot {
+
+      private LevelsSnapshot {
+        levels = levels.stream().toList(); //make it unmodifiable
+      }
+
+    }
+
+    @Override
+    public InputValidation validate() {
+      //valid if at least one level is present
+      InputValidation validation =
+          isEmpty() ? InputValidation.failed() : InputValidation.passed();
+      if (!validation.hasPassed()) {
+        setErrorMessage("Please provide at least one level.");
+      }
+      setInvalid(!validation.hasPassed());
+      return validation;
+    }
+
+    public List<LevelChange> getChanges() {
+      if (initialState instanceof LevelsSnapshot(List<Snapshot> levels)) {
+
+        var previousLevels = extractFilledLevels().apply(
+            levels
+                .stream()
+                .map(this::mapToRestoredLevelField)
+                .toList()
+        );
+        var currentLevels = extractFilledLevels().apply(levelsContainer.stream().toList());
+        //find all the deletions
+        List<LevelDeleted> levelDeletions = previousLevels.stream()
+            .filter(previousLevel -> !currentLevels.contains(previousLevel))
+            .map(removedLevel -> new LevelDeleted(previousLevels.indexOf(removedLevel)))
+            .toList();
+        // find all additions
+        List<LevelAdded> levelAdditions = currentLevels.stream()
+            .filter(currentLevel -> !previousLevels.contains(currentLevel))
+            .map(addedLevel -> new LevelAdded(currentLevels.indexOf(addedLevel), addedLevel))
+            .toList();
+        //find all moved levels
+        List<LevelMoved> levelMoves = previousLevels.stream()
+            .filter(currentLevels::contains)
+            .filter(previousLevel -> previousLevels.indexOf(previousLevel) != currentLevels.indexOf(
+                previousLevel))
+            .map(movedLevel -> new LevelMoved(previousLevels.indexOf(movedLevel),
+                currentLevels.indexOf(movedLevel)))
+            .toList();
+        var changes = new ArrayList<LevelChange>();
+        changes.addAll(levelDeletions);
+        changes.addAll(levelAdditions);
+        changes.addAll(levelMoves);
+        return changes;
+      }
+
+      throw new IllegalStateException();
+    }
+
+    private LevelField mapToRestoredLevelField(Snapshot snapshot) {
+      var dummy = new LevelField();
+      try {
+        dummy.restore(snapshot);
+        return dummy;
+      } catch (SnapshotRestorationException it) {
+        throw new IllegalStateException("Initial state not correctly set", it);
+      }
+    }
+
+    @Override
+    public boolean hasChanges() {
+      return levelsContainer.stream().anyMatch(LevelField::hasChanges);
+    }
+
+
+    public void markInitialized() {
+      levelsContainer.stream().forEach(LevelField::markInitialized);
+      this.initialState = snapshot();
+    }
+
+    @Override
+    public Snapshot snapshot() {
+      return new LevelsSnapshot(levelsContainer.stream().map(LevelField::snapshot).toList());
+    }
+
+    @Override
+    public void restore(Snapshot snapshot) throws SnapshotRestorationException {
+      if (snapshot instanceof LevelsSnapshot(List<Snapshot> levels)) {
+        List<LevelField> previousLevels = levelsContainer.stream().toList();
+        levelsContainer.clear();
+        try {
+          levels.stream()
+              .map(this::mapToRestoredLevelField)
+              .forEach(this::addLevel);
+          return;
+        } catch (Exception e) {
+          levelsContainer.clear();
+          for (LevelField previousLevel : previousLevels) {
+            levelsContainer.add(previousLevel);
+          }
+          throw e;
+        }
+      }
+      throw new SnapshotRestorationException(
+          "Unknown snapshot type. Expected %s but got %s".formatted(LevelsSnapshot.class,
+              snapshot.getClass()));
+
+    }
+  }
+
+  static class LevelField extends Composite<Div> implements Focusable<Component>,
+      CanSnapshot,
+      UserInput,
+      HasValidationProperties {
+
+
+    static final String LEVEL_CLASS = "level";
+    static final String LEVEL_FIELD_CSS = "flex-horizontal gap-03 width-full no-flex-wrap no-wrap input-with-label";
+    static final String LEVEL_VALUE_CSS = "dynamic-growing-flex-item";
+
+    private final TextField levelValue = new TextField();
+    private final Button deleteLevelButton;
+    private LevelSnapshot initialState;
+    private final SerializablePredicate<String> levelValidator = s -> !s.isBlank();
+
+    private LevelField() {
+      var buttonFactory = new ButtonFactory();
+      this.deleteLevelButton = buttonFactory.createGreyIconButton(VaadinIcon.TRASH.create());
+      deleteLevelButton.addClickListener(
+          clickEvent -> fireEvent(new DeleteLevelEvent(this, clickEvent.isFromClient())));
+      initialState = (LevelSnapshot) this.snapshot();
+    }
+
+    @Override
+    public void setErrorMessage(String errorMessage) {
+      levelValue.setErrorMessage(errorMessage);
+    }
+
+    @Override
+    public String getErrorMessage() {
+      return levelValue.getErrorMessage();
+    }
+
+    @Override
+    public void setInvalid(boolean invalid) {
+      levelValue.setInvalid(invalid);
+    }
+
+    @Override
+    public boolean isInvalid() {
+      return levelValue.isInvalid();
+    }
+
+    public Registration addValueChangeListener(
+        ValueChangeListener<? super ComponentValueChangeEvent<TextField, String>> listener) {
+      return levelValue.addValueChangeListener(listener);
+    }
+
+
+    @Override
+    protected Div initContent() {
+      var levelField = new Div();
+      levelField.addClassNames(LEVEL_FIELD_CSS);
+      levelValue.addClassNames(LEVEL_VALUE_CSS);
+      levelField.add(levelValue, deleteLevelButton);
+      levelField.addClassNames(LEVEL_CLASS);
+      return levelField;
+    }
+
+    /**
+     * After modification, this method should be used to inform the level field that its current
+     * value is to be seen as initial value of the field.
+     */
+    public void markInitialized() {
+      initialState = (LevelSnapshot) this.snapshot();
+    }
+
+    @Override
+    public void focus() {
+      levelValue.focus();
+    }
+
+    @Override
+    public void setTabIndex(int tabIndex) {
+      levelValue.setTabIndex(tabIndex);
+    }
+
+    @Override
+    public int getTabIndex() {
+      return levelValue.getTabIndex();
+    }
+
+    @Override
+    public void blur() {
+      levelValue.blur();
+    }
+
+    @Override
+    public ShortcutRegistration addFocusShortcut(Key key, KeyModifier... keyModifiers) {
+      return levelValue.addFocusShortcut(key, keyModifiers);
+    }
+
+    @Override
+    public Registration addFocusListener(ComponentEventListener<FocusEvent<Component>> listener) {
+      return levelValue.addFocusListener(
+          it -> listener.onComponentEvent(new FocusEvent<>(it.getSource(), it.isFromClient())));
+    }
+
+    public boolean isEmpty() {
+      return levelValue.isEmpty() || levelValue.getValue().isBlank();
+    }
+
+    public boolean isDeletion() {
+      return isEmpty() && !initialState.levelValue().equals(levelValue.getEmptyValue());
+    }
+
+    public void setValue(String value) {
+      levelValue.setValue(value);
+    }
+
+    public Optional<String> getValue() {
+      return levelValue.getOptionalValue().map(String::trim);
+    }
+
+    Registration addDeleteListener(ComponentEventListener<DeleteLevelEvent> listener) {
+      return addListener(DeleteLevelEvent.class, listener);
+    }
+
+    public void lock() {
+      deleteLevelButton.setEnabled(false);
+      deleteLevelButton.setVisible(false);
+      levelValue.setSuffixComponent(VaadinIcon.LOCK.create());
+      levelValue.setReadOnly(true);
+    }
+
+    record LevelSnapshot(@NonNull String levelValue) implements
+        Snapshot,
+        Serializable {
+
+      LevelSnapshot {
+        Objects.requireNonNull(levelValue);
+      }
+
+    }
+
+    @Override
+    public Snapshot snapshot() {
+      return new LevelSnapshot(levelValue.getValue());
+    }
+
+    @Override
+    public void restore(Snapshot snapshot) throws SnapshotRestorationException {
+      if (snapshot instanceof LevelSnapshot(String value)) {
+        levelValue.setValue(value);
         return;
       }
-      this.unitField.setValue(unit);
-    }
-
-    private Optional<String> readUnitValue() {
-      if (unitField.isEmpty()) {
-        return Optional.empty();
-      }
-      return Optional.of(unitField.getValue());
-    }
-
-    @Override
-    protected void setPresentationValue(ExperimentalVariableInformation newPresentationValue) {
-      if (isNull(newPresentationValue)) {
-        nameField.clear();
-        unitField.clear();
-        levelArea.setLevels(List.of());
-        return;
-      }
-      nameField.setValue(newPresentationValue.variableName());
-      setUnitValue(newPresentationValue.unit());
-      levelArea.setLevels(newPresentationValue.levels()); //FIXME preserve locking
-    }
-
-
-
-    public void setRemoveButtonListener(
-        Consumer<RemoveButtonClickEvent> closeListener) {
-      if (nonNull(clickListener)) {
-        clickListener.remove();
-      }
-      clickListener = deleteIcon.addClickListener(it -> closeListener.accept(
-          new RemoveButtonClickEvent(this)));
+      throw new SnapshotRestorationException(
+          "Unknown snapshot type. Expected %s but got %s".formatted(LevelSnapshot.class,
+              snapshot.getClass()));
     }
 
     @Override
     @NonNull
     public InputValidation validate() {
-      if (isEmpty()) {
-        return passed();
+      if (levelValidator.test(levelValue.getValue())) {
+        return InputValidation.passed();
       }
-      if (nameField.isRequired() && nameField.isEmpty()) {
-        nameField.setInvalid(true);
-      }
-      if (unitField.isRequired() && unitField.isEmpty()) {
-        unitField.setInvalid(true);
-      }
-      if (nameField.isInvalid() || unitField.isInvalid() || levelArea.isEmpty()) {
-        setInvalid(true);
-        return failed();
-      }
-      setInvalid(false);
-      return passed();
+      return InputValidation.failed();
     }
 
     @Override
     public boolean hasChanges() {
-      if (isNull(initialValue) && !isEmpty()) {
-        //something was filled in where nothing was before
-        return true;
+      return !initialState.levelValue().equals(levelValue.getValue());
+    }
+
+    @Override
+    public String toString() {
+      return new StringJoiner(", ", LevelField.class.getSimpleName() + "[", "]")
+          .add("levelValue=" + levelValue.getValue())
+          .add("hashCode=" + hashCode())
+          .toString();
+    }
+
+    static class DeleteLevelEvent extends ComponentEvent<LevelField> {
+
+      public DeleteLevelEvent(LevelField source, boolean fromClient) {
+        super(source, fromClient);
       }
-      return getOptionalValue().map(it -> !it.equals(initialValue)).orElse(false);
     }
-
-    public record RemoveButtonClickEvent(ExperimentalVariableRow origin) {
-
-    }
-
   }
+
+  @DomEvent(value = "paste",
+      allowUpdates = DisabledUpdateMode.ONLY_WHEN_ENABLED,
+      preventDefault = true,
+      stopPropagation = true)
+  public static class PasteEvent extends ComponentEvent<Component> {
+
+    private final String clipboardText;
+
+    /**
+     * Creates a new event using the given source and indicator whether the event originated from
+     * the client side or the server side.
+     *
+     * @param source     the source component
+     * @param fromClient <code>true</code> if the event originated from the client
+     *                   side, <code>false</code> otherwise
+     */
+    public PasteEvent(Component source, boolean fromClient,
+        @EventData("event.clipboardData.getData(\"text\")") String clipboardText) {
+      super(source, fromClient);
+      this.clipboardText = clipboardText;
+    }
+
+    public String getClipboardText() {
+      return clipboardText;
+    }
+  }
+
+
 }
